Create a Raster Band Object

Creates a band object for use with the Raster Extension. Describes the characteristics of a single raster band including data type, nodata values, scale/offset transforms, and statistics.

Usage

raster_band(
  nodata = NULL,
  data_type = NULL,
  unit = NULL,
  statistics = NULL,
  sampling = NULL,
  bits_per_sample = NULL,
  spatial_resolution = NULL,
  scale = NULL,
  offset = NULL,
  histogram = NULL,
  ...
)

Arguments

nodata

(numeric or NULL, optional) Pixel value(s) that should be interpreted as "no data". Can be a single value or vector of values. Common values: 0, -9999, NaN.

data_type

(character, optional) Data type of the band. Must be one of: "int8", "int16", "int32", "int64", "uint8", "uint16", "uint32", "uint64", "float16", "float32", "float64", "cint16", "cint32", "cfloat32", "cfloat64", or "other".

unit

(character, optional) Unit of measurement for the pixel values. Examples: "m" (meters), "W⋅sr⁻¹⋅m⁻²" (radiance), "1" (unitless/reflectance).

statistics

(list, optional) Statistics object created with raster_statistics() describing the distribution of pixel values.

sampling

(character, optional) Pixel sampling method. Either "area" (pixel represents area) or "point" (pixel represents point). Default is "area".

bits_per_sample

(integer, optional) Actual number of bits used for this band. Only needed when different from the standard for the data type (e.g., 1-bit data stored in uint8).

spatial_resolution

(numeric, optional) Average spatial resolution of pixels in the band, in meters. Useful when resolution varies or differs from ground sample distance (gsd).

scale

(numeric, optional) Multiplicative scaling factor to transform pixel values: physical_value = scale × DN + offset. Default is 1.

offset

(numeric, optional) Additive offset to transform pixel values: physical_value = scale × DN + offset. Default is 0.

histogram

(list, optional) Histogram object created with raster_histogram() describing the distribution of pixel values.

...

Additional fields for the band object. Can include fields from other extensions like "eo:common_name", "eo:center_wavelength".

Value

A list representing a raster band object.

Examples

# Simple band with just data type
band <- raster_band(data_type = "uint8")

# Band with nodata and scale/offset
band <- raster_band(
  nodata = 0,
  data_type = "uint16",
  scale = 0.0001,
  offset = 0,
  unit = "1"
)

# Band with statistics
band <- raster_band(
  nodata = -9999,
  data_type = "int16",
  spatial_resolution = 30,
  statistics = raster_statistics(
    minimum = -1000,
    maximum = 8000,
    mean = 2500,
    stddev = 1500
  )
)

# Band combining raster and EO extension
band <- raster_band(
  nodata = 0,
  data_type = "uint16",
  scale = 0.0001,
  offset = -0.1,
  spatial_resolution = 10,
  unit = "1",
  "eo:common_name" = "red",
  "eo:center_wavelength" = 0.665,
  "eo:full_width_half_max" = 0.038
)